controlla-me/trading/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "controlla-me-trading"
version = "0.1.0"
description = "Automated swing trading for Controlla.me financial sustainability"
requires-python = ">=3.11"
license = {text = "Private"}

dependencies = [
    # Broker
    "alpaca-py>=0.28.0",

    # Data & Analysis
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "ta>=0.11.0",             # Technical analysis indicators (RSI, MACD, Bollinger, etc.)

    # Database
    "supabase>=2.9.0",

    # Config & Validation
    "pydantic>=2.9.0",
    "pydantic-settings>=2.6.0",
    "python-dotenv>=1.0.0",

    # Scheduling
    "schedule>=1.2.0",

    # Logging
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pandas-stubs>=2.2.0",
]
backtest = [
    "matplotlib>=3.9.0",
    "pyarrow>=18.0.0",        # Parquet support
]

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM", "T20"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
